pyproject.toml
[project]
name = "sem-image-colorization"
version = "0.1.0"
description = "SEM image colorization project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair==5.3.0",
    "annotated-types==0.7.0",
    "attrs==23.2.0",
    "blinker==1.8.2",
    "blis==0.7.11",
    "cachetools==5.3.3",
    "catalogue==2.0.10",
    "certifi==2024.6.2",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cloudpathlib==0.18.1",
    "confection==0.1.5",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "cymem==2.0.8",
    "data==0.4",
    "dotenv>=0.9.9",
    "fastai==2.7.15",
    "fastapi>=0.117.1",
    "fastcore==1.5.47",
    "fastdownload==0.0.7",
    "fastprogress==1.0.3",
    "filelock==3.13.1",
    "fonttools==4.53.0",
    "fsspec==2024.2.0",
    "funcsigs==1.0.2",
    "future==1.0.0",
    "gitdb==4.0.11",
    "gitpython==3.1.43",
    "graphviz==0.20.3",
    "hiddenlayer==0.3",
    "idna==3.7",
    "imageio==2.34.1",
    "jinja2==3.1.3",
    "joblib==1.4.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2023.12.1",
    "kiwisolver==1.4.5",
    "langcodes==3.4.0",
    "language-data==1.2.0",
    "latex==0.7.0",
    "lazy-loader==0.4",
    "marisa-trie==1.2.0",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.9.0",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "murmurhash==1.0.10",
    "networkx==3.2.1",
    "numpy==1.26.3",
    "nvidia-cublas-cu11==11.11.3.6",
    "nvidia-cuda-cupti-cu11==11.8.87",
    "nvidia-cuda-nvrtc-cu11==11.8.89",
    "nvidia-cuda-runtime-cu11==11.8.89",
    "nvidia-cudnn-cu11==8.7.0.84",
    "nvidia-cufft-cu11==10.9.0.58",
    "nvidia-curand-cu11==10.3.0.86",
    "nvidia-cusolver-cu11==11.4.1.48",
    "nvidia-cusparse-cu11==11.7.5.86",
    "nvidia-nccl-cu11==2.20.5",
    "nvidia-nvtx-cu11==11.8.86",
    "opencv-python==4.10.0.84",
    "pandas==2.2.2",
    "pillow==10.2.0",
    "preshed==3.0.9",
    "protobuf==5.27.2",
    "pyarrow==16.1.0",
    "pydantic==2.7.4",
    "pydantic-core==2.18.4",
    "pydeck==0.9.1",
    "pyparsing==3.1.2",
    "pyqt5-qt5==5.15.14",
    "pyqt5-sip==12.15.0",
    "pyqt6-qt6==6.7.2",
    "pyqt6-sip==13.8.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rich==13.7.1",
    "rpds-py==0.19.0",
    "scienceplots==2.1.1",
    "scikit-image==0.24.0",
    "scikit-learn==1.5.0",
    "scipy==1.13.1",
    "setuptools==69.5.1",
    "shellingham==1.5.4",
    "shutilwhich==1.1.0",
    "smart-open==7.0.4",
    "smmap==5.0.1",
    "spacy==3.7.5",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.4.8",
    "stqdm==0.0.5",
    "streamlit==1.36.0",
    "streamlit-image-comparison==0.0.4",
    "streamlit-js-eval==0.1.7",
    "sympy==1.12",
    "tempdir==0.7.1",
    "tenacity==8.5.0",
    "thinc==8.2.5",
    "threadpoolctl==3.5.0",
    "tifffile==2024.6.18",
    "toml==0.10.2",
    "toolz==0.12.1",
    "torch==2.3.0",
    "torchsummary==1.5.1",
    "torchvision==0.18.0",
    "torchviz==0.0.2",
    "tqdm==4.66.4",
    "typer==0.12.3",
    "typing-extensions==4.9.0",
    "tzdata==2024.1",
    "urllib3==2.2.2",
    "uvicorn>=0.37.0",
    "wasabi==1.1.3",
    "watchdog==4.0.1",
    "weasel==0.4.1",
    "wheel==0.43.0",
    "wrapt==1.16.0",
]
